ann_requirement.txt
numpy
pandas
pillow
streamlit
tensorflow
scikit-learn
